EC_Central/requirements.txt
fastapi
uvicorn
kafka-python
pyjwt
cryptography
python-dotenv
